requirements.txt
asgiref
astroid
beautifulsoup4
bson
certifi
chardet
colorama
dataclasses
dj-database-url
Django
django-extensions
django-heroku
docopt
gunicorn
heroku
idna
isort
lazy-object-proxy
mccabe
psycopg2
pytz
requests
rope
six
soupsieve
sqlparse
urllib3
whitenoise
wincertstore
wrapt
<<<<<<< HEAD
=======
django-cors-headers
djangorestframework
>>>>>>> vue
